light: move HDR_PackE5BRG9/HDR_UnpackE5BRG9 to common/litfile.hh
